separated card tests based on sets, fixes #24 reconnecting bug, fixes #23
